fix(e2e): restore tracing checks
